MERGE: -c 29777 /trunk/Sapphire -> /branches/lwidentity-5.0
Fix potential build break on solaris.

(r29778)

+++ lsass/nsswitch/common/nss-netgrp.c
@@ -261,7 +261,7 @@ LsaNssCommonNetgroupParse(
             {
                 (*ppszCursor)++;
             }
-            else if (**ppszCursor != '(' && !isalpha(**ppszCursor))
+            else if (**ppszCursor != '(' && !isalpha((int)**ppszCursor))
             {
                 ret = NSS_STATUS_UNAVAIL;
                 BAIL_ON_NSS_ERROR(ret);
